introduce die-after option
Dice will kill itself once it reached that uptime.

There are issues where the watchers sometimes stop listening with the
following error:

```
ERROR: logging before flag.Parse: W0504 23:22:10.871644      1 reflector.go:270] k8s.io/client-go/informers/factory.go:131: watch of *v1.Pod ended with: too old resource version: 73937810 (73939460)
```

Unfortunately, this is stopped on the kubernetes side of the controller,
and we can't catch it. So when it happens, we don't listen anymore.

With this new option, we can configure dice to kill itself a bit after
the max uptime is reached. That way, if we're not listening anymore, we
will not remain blocked forever.